+++ clients/admin-ui/src/features/data-discovery-and-detection/tables/cells/DiscoveryItemActionsCell.tsx
@@ -4,17 +4,17 @@ import { useAlert } from "~/features/common/hooks";
 import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
 import { DiffStatus } from "~/types/api";
 
-import ActionButton from "./ActionButton";
+import ActionButton from "../../ActionButton";
 import {
   useMuteResourceMutation,
   usePromoteResourceMutation,
-} from "./discovery-detection.slice";
+} from "../../discovery-detection.slice";
 
 interface DiscoveryItemActionsProps {
   resource: DiscoveryMonitorItem;
 }
 
-const DiscoveryItemActions = ({ resource }: DiscoveryItemActionsProps) => {
+const DiscoveryItemActionsCell = ({ resource }: DiscoveryItemActionsProps) => {
   const [promoteResourceMutation, { isLoading: promoteIsLoading }] =
     usePromoteResourceMutation();
   const [muteResourceMutation, { isLoading: muteIsLoading }] =
@@ -89,4 +89,4 @@ const DiscoveryItemActions = ({ resource }: DiscoveryItemActionsProps) => {
   );
 };
 
-export default DiscoveryItemActions;
+export default DiscoveryItemActionsCell;