+++ clients/admin-ui/src/features/data-discovery-and-detection/hooks/useDetectionResultColumns.tsx
@@ -2,11 +2,12 @@ import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 
 import { DefaultCell, DefaultHeaderCell } from "~/features/common/table/v2";
 import { RelativeTimestampCell } from "~/features/common/table/v2/cells";
-import DetectionItemAction from "~/features/data-discovery-and-detection/DetectionItemActions";
-import ResultStatusBadgeCell from "~/features/data-discovery-and-detection/tables/ResultStatusBadgeCell";
-import ResultStatusCell from "~/features/data-discovery-and-detection/tables/ResultStatusCell";
+import DetectionItemActionsCell from "~/features/data-discovery-and-detection/tables/cells/DetectionItemActionsCell";
+import FieldDataTypeCell from "~/features/data-discovery-and-detection/tables/cells/FieldDataTypeCell";
+import ResultStatusBadgeCell from "~/features/data-discovery-and-detection/tables/cells/ResultStatusBadgeCell";
+import ResultStatusCell from "~/features/data-discovery-and-detection/tables/cells/ResultStatusCell";
+import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
 import { StagedResourceType } from "~/features/data-discovery-and-detection/types/StagedResourceType";
-import { StagedResourceAPIResponse } from "~/types/api";
 
 import findProjectFromUrn from "../utils/findProjectFromUrn";
 
@@ -15,9 +16,9 @@ const useDetectionResultColumns = ({
 }: {
   resourceType?: StagedResourceType;
 }) => {
-  const columnHelper = createColumnHelper<StagedResourceAPIResponse>();
+  const columnHelper = createColumnHelper<DiscoveryMonitorItem>();
 
-  const defaultColumns: ColumnDef<StagedResourceAPIResponse, any>[] = [];
+  const defaultColumns: ColumnDef<DiscoveryMonitorItem, any>[] = [];
 
   if (!resourceType) {
     return { columns: defaultColumns };
@@ -59,7 +60,9 @@ const useDetectionResultColumns = ({
       }),
       columnHelper.display({
         id: "actions",
-        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        cell: (props) => (
+          <DetectionItemActionsCell resource={props.row.original} />
+        ),
         header: "Actions",
       }),
     ];
@@ -75,7 +78,7 @@ const useDetectionResultColumns = ({
       }),
       columnHelper.accessor((row) => row.description, {
         id: "description",
-        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        cell: (props) => <DefaultCell value={props.getValue()} />,
         header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
@@ -95,7 +98,9 @@ const useDetectionResultColumns = ({
       }),
       columnHelper.display({
         id: "actions",
-        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        cell: (props) => (
+          <DetectionItemActionsCell resource={props.row.original} />
+        ),
         header: "Actions",
       }),
     ];
@@ -109,9 +114,14 @@ const useDetectionResultColumns = ({
         cell: (props) => <ResultStatusCell result={props.row.original} />,
         header: (props) => <DefaultHeaderCell value="Field name" {...props} />,
       }),
+      columnHelper.accessor((row) => row.source_data_type, {
+        id: "data-type",
+        cell: (props) => <FieldDataTypeCell type={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Data type" {...props} />,
+      }),
       columnHelper.accessor((row) => row.description, {
         id: "description",
-        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        cell: (props) => <DefaultCell value={props.getValue()} />,
         header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
@@ -131,7 +141,9 @@ const useDetectionResultColumns = ({
       }),
       columnHelper.display({
         id: "actions",
-        cell: (props) => <DetectionItemAction resource={props.row.original} />,
+        cell: (props) => (
+          <DetectionItemActionsCell resource={props.row.original} />
+        ),
         header: "Actions",
       }),
     ];