+++ clients/admin-ui/src/features/data-discovery-and-detection/hooks/useDiscoveryResultColumns.tsx
@@ -5,24 +5,26 @@ import {
   IndeterminateCheckboxCell,
   RelativeTimestampCell,
 } from "~/features/common/table/v2/cells";
-import ResultStatusBadgeCell from "~/features/data-discovery-and-detection/tables/ResultStatusBadgeCell";
+import FieldDataTypeCell from "~/features/data-discovery-and-detection/tables/cells/FieldDataTypeCell";
+import ResultStatusBadgeCell from "~/features/data-discovery-and-detection/tables/cells/ResultStatusBadgeCell";
+import { DiscoveryMonitorItem } from "~/features/data-discovery-and-detection/types/DiscoveryMonitorItem";
 import { ResourceChangeType } from "~/features/data-discovery-and-detection/types/ResourceChangeType";
 import { StagedResourceType } from "~/features/data-discovery-and-detection/types/StagedResourceType";
 import findProjectFromUrn from "~/features/data-discovery-and-detection/utils/findProjectFromUrn";
-import { DiffStatus, StagedResourceAPIResponse } from "~/types/api";
+import { DiffStatus } from "~/types/api";
 
-import DiscoveryItemActions from "../DiscoveryItemActions";
-import ResultStatusCell from "../tables/ResultStatusCell";
-import TaxonomyDisplayAndEdit from "../TaxonomyDisplayAndEdit";
+import DiscoveryItemActionsCell from "../tables/cells/DiscoveryItemActionsCell";
+import EditCategoriesCell from "../tables/cells/EditCategoryCell";
+import ResultStatusCell from "../tables/cells/ResultStatusCell";
 
 const useDiscoveryResultColumns = ({
   resourceType,
 }: {
   resourceType: StagedResourceType | undefined;
 }) => {
-  const columnHelper = createColumnHelper<StagedResourceAPIResponse>();
+  const columnHelper = createColumnHelper<DiscoveryMonitorItem>();
 
-  const defaultColumns: ColumnDef<StagedResourceAPIResponse, any>[] = [];
+  const defaultColumns: ColumnDef<DiscoveryMonitorItem, any>[] = [];
 
   if (resourceType === StagedResourceType.SCHEMA) {
     const columns = [
@@ -67,7 +69,7 @@ const useDiscoveryResultColumns = ({
         id: "action",
         cell: (props) =>
           props.row.original.diff_status !== DiffStatus.MUTED ? (
-            <DiscoveryItemActions resource={props.row.original} />
+            <DiscoveryItemActionsCell resource={props.row.original} />
           ) : (
             <DefaultCell value="--" />
           ),
@@ -105,7 +107,7 @@ const useDiscoveryResultColumns = ({
       }),
       columnHelper.accessor((row) => row.description, {
         id: "description",
-        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        cell: (props) => <DefaultCell value={props.getValue()} />,
         header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
@@ -125,7 +127,9 @@ const useDiscoveryResultColumns = ({
       }),
       columnHelper.display({
         id: "actions",
-        cell: (props) => <DiscoveryItemActions resource={props.row.original} />,
+        cell: (props) => (
+          <DiscoveryItemActionsCell resource={props.row.original} />
+        ),
         header: "Actions",
       }),
     ];
@@ -139,9 +143,14 @@ const useDiscoveryResultColumns = ({
         cell: (props) => <ResultStatusCell result={props.row.original} />,
         header: (props) => <DefaultHeaderCell value="Field name" {...props} />,
       }),
+      columnHelper.accessor((row) => row.source_data_type, {
+        id: "data-type",
+        cell: (props) => <FieldDataTypeCell type={props.getValue()} />,
+        header: (props) => <DefaultHeaderCell value="Data type" {...props} />,
+      }),
       columnHelper.accessor((row) => row.description, {
         id: "description",
-        cell: (props) => <DefaultCell value={props.getValue() ?? "--"} />,
+        cell: (props) => <DefaultCell value={props.getValue()} />,
         header: (props) => <DefaultHeaderCell value="Description" {...props} />,
       }),
       columnHelper.display({
@@ -157,7 +166,7 @@ const useDiscoveryResultColumns = ({
       columnHelper.display({
         id: "classifications",
         cell: ({ row }) => {
-          return <TaxonomyDisplayAndEdit resource={row.original} />;
+          return <EditCategoriesCell resource={row.original} />;
         },
         meta: { overflow: "visible", disableRowClick: true },
         header: "Data category",
@@ -170,7 +179,9 @@ const useDiscoveryResultColumns = ({
       }),
       columnHelper.display({
         id: "actions",
-        cell: (props) => <DiscoveryItemActions resource={props.row.original} />,
+        cell: (props) => (
+          <DiscoveryItemActionsCell resource={props.row.original} />
+        ),
         header: "Actions",
       }),
     ];